# specs/002-backend-api-database/plan/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Todo API
  description: Backend API for managing user tasks with JWT authentication
  version: 1.0.0
servers:
  - url: http://localhost:8000
    description: Development server
  - url: https:/api.todoapp.com
    description: Production server
paths:
  /api/tasks:
    get:
      summary: Get all tasks for authenticated user
      description: Retrieve a list of all tasks owned by the authenticated user
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Successfully retrieved user's tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Task'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '500':
          $ref: '#/components/responses/InternalServerError'
    post:
      summary: Create a new task
      description: Create a new task associated with the authenticated user
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTaskRequest'
      responses:
        '201':
          description: Task created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '422':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/tasks/{task_id}:
    get:
      summary: Get a specific task
      description: Retrieve details of a specific task owned by the authenticated user
      security:
        - BearerAuth: []
      parameters:
        - name: task_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: The UUID of the task to retrieve
      responses:
        '200':
          description: Successfully retrieved task
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/InternalServerError'
    put:
      summary: Update a specific task
      description: Update details of a specific task owned by the authenticated user
      security:
        - BearerAuth: []
      parameters:
        - name: task_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: The UUID of the task to update
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTaskRequest'
      responses:
        '200':
          description: Task updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '422':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/InternalServerError'
    delete:
      summary: Delete a specific task
      description: Delete a specific task owned by the authenticated user
      security:
        - BearerAuth: []
      parameters:
        - name: task_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: The UUID of the task to delete
      responses:
        '204':
          description: Task deleted successfully
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/InternalServerError'
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Task:
      type: object
      required:
        - id
        - title
        - user_id
        - status
        - created_at
        - updated_at
      properties:
        id:
          type: string
          format: uuid
          description: Unique identifier for the task
        title:
          type: string
          maxLength: 255
          description: Title of the task
        description:
          type: string
          nullable: true
          description: Detailed description of the task
        status:
          type: string
          enum: [pending, in-progress, completed]
          default: pending
          description: Current status of the task
        user_id:
          type: string
          format: uuid
          description: ID of the user who owns this task
        created_at:
          type: string
          format: date-time
          description: Timestamp when the task was created
        updated_at:
          type: string
          format: date-time
          description: Timestamp when the task was last updated
    CreateTaskRequest:
      type: object
      required:
        - title
      properties:
        title:
          type: string
          maxLength: 255
          description: Title of the task
        description:
          type: string
          nullable: true
          description: Detailed description of the task
        status:
          type: string
          enum: [pending, in-progress, completed]
          default: pending
          description: Initial status of the task
    UpdateTaskRequest:
      type: object
      properties:
        title:
          type: string
          maxLength: 255
          description: Updated title of the task
        description:
          type: string
          nullable: true
          description: Updated description of the task
        status:
          type: string
          enum: [pending, in-progress, completed]
          description: Updated status of the task
  responses:
    UnauthorizedError:
      description: Unauthorized - Missing or invalid JWT token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            detail: "Not authenticated"
    NotFoundError:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            detail: "Task not found"
    BadRequestError:
      description: Bad request - Invalid input data
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            detail: "Invalid input data"
    ValidationError:
      description: Validation error - Request failed validation
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            detail: "Validation error"
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            detail: "Internal server error"
    Error:
      type: object
      required:
        - detail
      properties:
        detail:
          type: string
          description: Error message